chore: move go router extension into other folder
