+++ lib/app/extensions/extensions.dart
@@ -3,6 +3,7 @@ export 'bool.dart';
 export 'build_context.dart';
 export 'color.dart';
 export 'enum.dart';
+export 'go_router_state.dart';
 export 'iterable.dart';
 export 'list.dart';
 export 'num.dart';
